Sync up to changes and update actions
